source/requirements.txt
simplematrixbotlib
python-a2s
rcon